+++ srcs/scene/elements/cylinder.h
@@ -5,17 +5,17 @@
 typedef struct s_cylinder
 {
     // x,y,z coordinates of the center of the cylinder.
-    float   *coordinates;
+    double   *coordinates;
 
     // 3d normalized vector of axis of cylinder. In range [-1,1],
     // for each x,y,z axis.
-    float   *vector;
+    double   *vector;
 
     // The cylinder diameter.
-    float   *diameter;
+    double   *diameter;
 
     // The cylinder height.
-    float   *height;
+    double   *height;
 
     // R,G,B cylinder colors in range [0,255].
     int     *color;
