+++ srcs/scene/elements/sphere.h
@@ -5,10 +5,10 @@
 typedef struct s_sphere
 {
     // x,y,z coordinates of the the sphere center.
-    float   *coordinates;
+    double   *coordinates;
 
     // The sphere diameter.
-    float   diameter;
+    double   diameter;
 
     // R,G,B sphere color in range [0-255].
     int     *color;
